backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.34.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-tinymce==3.6.1
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
Markdown==3.4.4
packaging==23.1
Pillow==10.0.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.5.0
